+++ client/views/__init__.py
@@ -1,0 +1,5 @@
+from .view import View
+from .menuview import MenuView
+from .loginview import LoginView
+from .registerview import RegisterView
+from .gameview import GameView
